refactor: Split loading and layout components
