+++ src/drive/web/modules/move/MoveFileList.jsx
@@ -1,0 +1,42 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import { DumbFile as File } from 'drive/web/modules/filelist/File'
+import LoadMore from 'drive/web/modules/filelist/LoadMore'
+
+const isValidMoveTarget = (subjects, target) => {
+  const isASubject = subjects.find(subject => subject._id === target._id)
+  const isAFile = target.type === 'file'
+
+  return isAFile || isASubject
+}
+
+const MoveFileList = ({ targets, files, hasMore, fetchMore, navigateTo }) => (
+  <div>
+    {files.map(file => (
+      <File
+        key={file.id}
+        disabled={isValidMoveTarget(targets, file)}
+        attributes={file}
+        displayedFolder={null}
+        actions={null}
+        isRenaming={false}
+        onFolderOpen={id => navigateTo(files.find(f => f.id === id))}
+        onFileOpen={null}
+        withSelectionCheckbox={false}
+        withFilePath={false}
+        withSharedBadge={true}
+      />
+    ))}
+    {hasMore && <LoadMore onClick={fetchMore} isLoading={false} />}
+  </div>
+)
+
+MoveFileList.propTypes = {
+  targets: PropTypes.array.isRequired,
+  files: PropTypes.array.isRequired,
+  hasMore: PropTypes.bool.isRequired,
+  fetchMore: PropTypes.func.isRequired,
+  navigateTo: PropTypes.func.isRequired
+}
+
+export default MoveFileList
